trigger not working
