+++ data/include/sensors.h
@@ -7,89 +7,24 @@
 #include "I2CGPS.h"
 #include "TinyGPS++.h"
 #include "adxl375.h"
+#include "data.h"
 
 
 
-struct SensStatus {
-    uint8_t     adxl_status;
-    uint8_t     adxl_aux_status;
-    uint8_t     bmi_accel_status;
-    uint8_t     bmi_aux_accel_status;
-    uint8_t     bmi_gyro_status;
-    uint8_t     bmi_aux_gyro_status;
-    bmp3_status bmp_status;
-    bmp3_status bmp_aux_status;
-};
-
-struct SensRaw {
-    adxl375_data        adxl;
-    adxl375_data        adxl_aux;
-    bmi08_sensor_data_f bmi_accel;
-    bmi08_sensor_data_f bmi_gyro; // Correspondance x,y,z / roll,pitch,yaw à déterminer
-    bmi08_sensor_data_f bmi_aux_accel;
-    bmi08_sensor_data_f bmi_aux_gyro;
-    bmp3_data           bmp;
-    bmp3_data           bmp_aux;
-
-    SensRaw();
-};
-
-struct Vector3 {
-    double x;
-    double y;
-    double z;
-};
-
-struct GPSCoord {
-    double lat;
-    double lng;
-    double alt;
-};
-
-struct GPSTime {
-    unsigned year;
-    unsigned month;
-    unsigned day;
-    unsigned hour;
-    unsigned minute;
-    unsigned second;
-    unsigned centisecond;
-};
-
-struct SensFiltered {
-    GPSTime   time;
-    GPSCoord  position;
-    Vector3   speed;
-    Vector3   accel;
-    Vector3   attitude;
-    double    course;
-    double    altitude;
-    bmp3_data baro;
-    double    N2_pressure;
-    double    fuel_pressure;
-    double    LOX_pressure;
-    double    fuel_level;
-    double    LOX_level;
-    double    engine_temperature;
-    double    igniter_pressure;
-    double    LOX_inj_pressure;
-    double    fuel_inj_pressure;
-    double    chamber_pressure;
-
-    SensFiltered();
-};
 
 class Sensors {
 public:
-    Sensors();
+    Sensors(Data data);
     ~Sensors();
 
     void calibrate();
     bool update();
+    bool write();
     inline SensStatus get_status() const { return status; }
     inline SensRaw get_raw() const { return raw_data; }
     inline SensFiltered get_clean() const { return clean_data; }
 private:
+    Data data;
     Adxl375 adxl1, adxl2;
     Bmi088 bmi1, bmi2;
     Bmp390 bmp1, bmp2;
