+++ data/src/data.cpp
@@ -6,7 +6,7 @@ Data::Data() {}
 Data::~Data() {}
 
 void Data::send() {
-    const SensFiltered sens(m_sensors.get_clean());
+    const SensFiltered sens = goat.sensors_data;
     av_downlink_t packet;
 
     packet.gnss_lat = sens.position.lat;
@@ -32,13 +32,7 @@ void Data::send() {
     m_telecom.send_packet(CAPSULE_ID::AV_TELEMETRY, (uint8_t*)&packet, av_downlink_size);
 }
 
-bool Data::update() {
-    m_telecom.update();
-    updated = m_sensors.update();
-
-    return updated;
-}
-
+//TODO mybe have a more granular dump
 DataDump Data::dump() const {
-    return {m_telecom.get_cmd(), m_sensors.get_raw(), m_sensors.get_clean()};
+    return {m_telecom.get_cmd(), goat.sensors_raw, goat.sensors_data};
 }
