+++ data/src/sensors.cpp
@@ -37,7 +37,7 @@ SensFiltered::SensFiltered()
     chamber_pressure(0)
 {}
 //TODO Where to import te defs (2 sources) ?
-Sensors::Sensors() 
+Sensors::Sensors(Data data)
 :   adxl1(ADXL375_ADDR_I2C_PRIM),
     adxl2(ADXL375_ADDR_I2C_SEC),
     bmi1(BMI08_ACCEL_I2C_ADDR_PRIMARY, BMI08_GYRO_I2C_ADDR_PRIMARY),
@@ -45,8 +45,9 @@ Sensors::Sensors()
     bmp1(BMP3_ADDR_I2C_PRIM),
     bmp2(BMP3_ADDR_I2C_SEC),
     i2cgps(),
-    gps() 
+    gps()
 {
+    this->data = data;
     update_status();
 }
 
@@ -59,58 +60,54 @@ void Sensors::calibrate() {
 
 bool Sensors::update() {
     update_status();
-
     // Update raw sensors values
-    raw_data.adxl = adxl1.get_data();
-    raw_data.adxl_aux = adxl2.get_data();
-
-    raw_data.bmi_accel = bmi1.get_accel_data();
-    raw_data.bmi_gyro = bmi1.get_gyro_data();
-    raw_data.bmi_aux_accel = bmi2.get_accel_data();
-    raw_data.bmi_aux_gyro = bmi2.get_gyro_data();
-
-    raw_data.bmp = bmp1.get_sensor_data();
-    raw_data.bmp_aux = bmp2.get_sensor_data();
-
+    this->data.goat.sensors_raw.adxl = adxl1.get_data();
+    this->data.goat.sensors_raw.adxl_aux = adxl2.get_data();
+    this->data.goat.sensors_raw.bmi_accel = bmi1.get_accel_data();
+    this->data.goat.sensors_raw.bmi_gyro = bmi1.get_gyro_data();
+    this->data.goat.sensors_raw.bmi_aux_accel = bmi2.get_accel_data();
+    this->data.goat.sensors_raw.bmi_aux_gyro = bmi2.get_gyro_data();
+    this->data.goat.sensors_raw.bmp = bmp1.get_sensor_data();
+    this->data.goat.sensors_raw.bmp_aux = bmp2.get_sensor_data();
     // Call Kalmann filter API
-    // Something like: clean_data = Kalmann(SensRaw raw_data); ?
+    // Something like: this->data.goat.sensors_data = Kalmann(SensRaw raw_data); ?
     
     while (i2cgps.available()) {
         gps.encode(i2cgps.read());
     }
     if (gps.time.isUpdated()) {
         if (gps.date.isValid()) {
-            clean_data.time.year = static_cast<int>(gps.date.year());
-            clean_data.time.month = static_cast<int>(gps.date.month());
-            clean_data.time.day = static_cast<int>(gps.date.day());
+            this->data.goat.sensors_data.time.year = static_cast<int>(gps.date.year());
+            this->data.goat.sensors_data.time.month = static_cast<int>(gps.date.month());
+            this->data.goat.sensors_data.time.day = static_cast<int>(gps.date.day());
         }
         if (gps.time.isValid()) {
-            clean_data.time.hour = static_cast<int>(gps.time.hour());
-            clean_data.time.minute = static_cast<int>(gps.time.minute());
-            clean_data.time.second = static_cast<int>(gps.time.second());
-            clean_data.time.centisecond = static_cast<int>(gps.time.centisecond());
+            this->data.goat.sensors_data.time.hour = static_cast<int>(gps.time.hour());
+            this->data.goat.sensors_data.time.minute = static_cast<int>(gps.time.minute());
+            this->data.goat.sensors_data.time.second = static_cast<int>(gps.time.second());
+            this->data.goat.sensors_data.time.centisecond = static_cast<int>(gps.time.centisecond());
         }
         if (gps.location.isValid()) {
-            clean_data.position.lat = gps.location.lat();
-            clean_data.position.lng = gps.location.lng();
+            this->data.goat.sensors_data.position.lat = gps.location.lat();
+            this->data.goat.sensors_data.position.lng = gps.location.lng();
         }
         if (gps.altitude.isValid()) {
-            clean_data.position.alt = gps.altitude.meters();
+            this->data.goat.sensors_data.position.alt = gps.altitude.meters();
         }
         if (gps.course.isValid()) {
-            clean_data.course = gps.course.deg();
+            this->data.goat.sensors_data.course = gps.course.deg();
         }
     }
     return true;
 }
 
 void Sensors::update_status() {
-    status.adxl_status = adxl1.get_status();
-    status.adxl_aux_status = adxl2.get_status();
-    status.bmi_accel_status = bmi1.get_accel_status();
-    status.bmi_aux_accel_status = bmi2.get_accel_status();
-    status.bmi_gyro_status = bmi1.get_gyro_status();
-    status.bmi_aux_gyro_status = bmi2.get_gyro_status();
-    status.bmp_status = bmp1.get_status();
-    status.bmp_aux_status = bmp2.get_status();
+    this->data.sensors_status.adxl_status = adxl1.get_status();
+    this->data.sensors_status.adxl_aux_status = adxl2.get_status();
+    this->data.sensors_status.bmi_accel_status = bmi1.get_accel_status();
+    this->data.sensors_status.bmi_aux_accel_status = bmi2.get_accel_status();
+    this->data.sensors_status.bmi_gyro_status = bmi1.get_gyro_status();
+    this->data.sensors_status.bmi_aux_gyro_status = bmi2.get_gyro_status();
+    this->data.sensors_status.bmp_status = bmp1.get_status();
+    this->data.sensors_status.bmp_aux_status = bmp2.get_status();
 }
